rename index test to plugin
